requirements.txt
numpy==1.16.3
scikit_image==0.15.0
Keras==2.2.4
opencv_python==4.1.0.25
joblib==0.13.2
tensorflow==1.13.1
matplotlib==3.1.0
six==1.11.0
Pillow==6.0.0
protobuf==3.8.0
object_detection==0.0.3
pymssql==2.1.4
scikit_learn==0.21.2
tqdm==4.32.1
